portal/requirements.txt
django==4.1.1
pillow==9.2.0
django-cleanup==6.0.0
django-filter==22.1
django-mptt==0.14.0
django-tinymce==3.5.0
django-debug-toolbar==3.7.0
django-sendfile2==0.7.0
python-dotenv==0.20.0
gunicorn==20.0.4
psycopg2-binary==2.9.3